third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/shadow/suppress-mutation-events-in-shadow-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Test that mutation events are not triggered in shadow DOM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutationEventTriggered is false
PASS successfullyParsed is true

TEST COMPLETE
